[Refactor] 베이스 다이얼로그 프래그먼트 적용 및 네이밍 변견
