+++ app/src/main/java/com/hara/kaera/presentation/dialog/DialogCompleteFragment.kt
@@ -4,11 +4,11 @@ import android.os.Bundle
 import android.view.View
 import androidx.core.widget.addTextChangedListener
 import com.hara.kaera.R
-import com.hara.kaera.databinding.DialogFragmentCompleteBinding
+import com.hara.kaera.databinding.DialogFragmentWorryMiningBinding
 import com.hara.kaera.presentation.base.BindingDialogFragment
 
 class DialogCompleteFragment :
-    BindingDialogFragment<DialogFragmentCompleteBinding>(R.layout.dialog_fragment_complete, 20) {
+    BindingDialogFragment<DialogFragmentWorryMiningBinding>(R.layout.dialog_fragment_worry_mining, 20) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
