+++ app/src/main/res/layout/dialog_fragment_worry_mining.xml
